Improve robustness when constructing roadSign object.
We're not sure whether the provided roadSign snipped does contain
e.g. board, opstelling etc... Hence the ".find" operation can return
undefined, and the subsequent ".key" call can cause an exception...

So that's why the "( find() || {}).key" construct.
However, this is not very readable, and should be rewritten a bit.
Consider this a TODO.